+++ src/northbridge/intel/nehalem/romstage.c
@@ -45,11 +45,6 @@ void mainboard_romstage_entry(void)
 	/* TODO, make this configurable */
 	nehalem_early_initialization(NEHALEM_MOBILE);
 
-	pch_pre_console_init();
-
-	/* Initialize console device(s) */
-	console_init();
-
 	early_pch_init();
 
 	/* Read PM1_CNT, DON'T CLEAR IT or raminit will fail! */
